backend/pyproject.toml
[project]
name = "weekend-plan-builder"
version = "0.1.0"
description = "Weekend Plan Builder - Agno agent with AG-UI interface"
requires-python = ">=3.11"
dependencies = [
    "agno[google]",
    "ag-ui-protocol",
    "ddgs",
    "uvicorn",
    "fastapi",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
